:truck:  The repository was divided into entities and types.
